サービスクラスを分解
